requirements.txt
aiogram==3.6.0
aiosqlite==0.20.0
python-dotenv==1.0.1
openai==1.40.6
APScheduler==3.10.4
pydantic==2.7.4